flex/csv-flex-template/requirements.txt
apache-beam[gcp]==2.64.0
google-cloud-bigquery==3.30.0
google-cloud-secret-manager==2.22.1
google-cloud-logging==3.11.4